Fixed uint->int problem when calculating MSE and PAE
Former-commit-id: a546a4edbb418f029a90017a61b5b404bd29bcc7

+++ enb/icompression.py
@@ -542,17 +542,17 @@ class LossyCompressionExperiment(CompressionExperiment):
     @atable.column_function("mse", label="MSE", plot_min=0)
     def set_MSE(self, index, row):
         original_array = np.fromfile(row.compression_results.original_path,
-                                     dtype=row.numpy_dtype)
+                                     dtype=row.numpy_dtype).astype(np.int64)
         reconstructed_array = np.fromfile(row.decompression_results.reconstructed_path,
-                                          dtype=row.numpy_dtype)
+                                          dtype=row.numpy_dtype).astype(np.int64)
         row[_column_name] = np.average((original_array - reconstructed_array) ** 2)
 
     @atable.column_function("pae", label="PAE", plot_min=0)
     def set_PAE(self, index, row):
         original_array = np.fromfile(row.compression_results.original_path,
-                                     dtype=row.numpy_dtype)
+                                     dtype=row.numpy_dtype).astype(np.int64)
         reconstructed_array = np.fromfile(row.decompression_results.reconstructed_path,
-                                          dtype=row.numpy_dtype)
+                                          dtype=row.numpy_dtype).astype(np.int64)
         row[_column_name] = np.max(np.abs(original_array - reconstructed_array))
